requirements.txt
webvtt-py==0.5.1
yt-dlp==2025.5.22
